+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 7.1.1
+
+* Bump Backendless Android SDK version to 6.3.1
+
+# 7.1.0
+
+Add:
+* `createEmailConfirmationURL` method
+* RT listeners for relations
+* Deep save
+
 # 7.0.0
 
 > Note: This release has breaking changes.